feat: use official planespotting api to fetch images
The third party API that we used was not properly implemented. It seemed
that it did a search on jetphotes based on the registration and simply
picked the first result. This did not always match the real aircraft.

Now we use the official API of planespotting.net. Here we can pass the
ICAO and always get the correct image if it exists.