fix: handle failure to get weatherdata
If the response of the weather API is not 200, we simply logged it but
did not throw an error. In a previous commit we handle the returned
error, but since we never actually returned it as an error, the problem
was not solved.

Now we effectively return an error if a non 200 status is returned, thus
the error handling function will take effect and properly handle it.