fix: use backend api
Instead of writing custom logic in our frontend to filter out the
aircraft without registrations, just call the backend instead.

This fixes the issue where the counter would stay 5 aircraft spotted,
while the UI only showed 3 for example. The other 2 would have no
registrations and were hidden from the UI, but not in the counter.