fix: split slack message if there are too many aircraft
Slack has a limit of 50 blocks per message. We use about 4 blocks per
aircraft. This means that if we spot 13 jets at the same time we get an
error when sending the slack message.

Now we have a maximum of 10 aircraft in each message. If there are
more aircraft they will be send in another message.