fix: remove calls to external api in frontend
Initially the frontend called a function that did a backend call to
fetch flight details information. This was done in the backend but also
in the frontend if you refreshed the page. This resulted in the posibility to hit the rate limit be spam refreshing
the dashboard page.

Now the backend is the only component doing the API calls to fetch
flight details. The frontend only displays it. As it should.